In progress refactor of Spring Security. Got rid of old StormpathWebSecurityConfigurer and renamed StormpathWebSecurityConfigurerAdapter to StormpathWebSecurityConfigurer. Refactored Spring Security WebMVC examples and is working. Refactored stormpath default spring boot starter and /login is NOT working.
