+++ examples/spring-boot-default/pom.xml
@@ -28,7 +28,7 @@
     <properties>
         <logback.version>1.1.2</logback.version>
         <slf4j.version>1.7.8</slf4j.version>
-        <spring.boot.version>1.2.5.RELEASE</spring.boot.version>
+        <spring.boot.version>1.3.0.RELEASE</spring.boot.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
