+++ examples/spring-security-spring-boot-webmvc/pom.xml
@@ -29,7 +29,7 @@
         <logback.version>1.1.2</logback.version>
         <slf4j.version>1.7.8</slf4j.version>
         <tomcat.version>7.0.59</tomcat.version>
-        <spring.boot.version>1.2.5.RELEASE</spring.boot.version>
+        <spring.boot.version>1.3.0.RELEASE</spring.boot.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
