+++ examples/spring-security-spring-boot-webmvc/src/main/java/com/stormpath/spring/boot/examples/SpringSecurityWebAppConfig.java
@@ -15,24 +15,29 @@
  */
 package com.stormpath.spring.boot.examples;
 
-import com.stormpath.spring.config.StormpathWebSecurityConfigurerAdapter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+import static com.stormpath.spring.config.StormpathWebSecurityConfigurer.stormpath;
 
 /**
  * @since 1.0.RC5
  */
 @Configuration
-public class SpringSecurityWebAppConfig extends StormpathWebSecurityConfigurerAdapter {
+public class SpringSecurityWebAppConfig extends WebSecurityConfigurerAdapter {
 
     /**
      * {@inheritDoc}
      */
     @Override
-    protected void doConfigure(HttpSecurity http) throws Exception {
+    protected void configure(HttpSecurity http) throws Exception {
         http
-                .authorizeRequests()
-                .antMatchers("/restricted").fullyAuthenticated();
+            .apply(stormpath())
+            .and()
+            .authorizeRequests()
+            .antMatchers("/restricted").fullyAuthenticated()
+            .antMatchers("/**").permitAll();
     }
 
 }
