+++ examples/spring-security-webmvc/src/main/java/com/stormpath/spring/examples/SpringSecurityWebAppConfig.java
@@ -21,7 +21,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import static com.stormpath.spring.config.StormpathWebSecurityConfigurerAdapter.*;
+import static com.stormpath.spring.config.StormpathWebSecurityConfigurer.*;
 
 /**
  * @since 1.0.RC6
