+++ extensions/spring/stormpath-spring-security-webmvc/src/main/java/com/stormpath/spring/config/StormpathWebSecurityConfigurer.java
@@ -1,18 +1,18 @@
 /*
-* Copyright 2015 Stormpath, Inc.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ * Copyright 2015 Stormpath, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.stormpath.spring.config;
 
 import com.stormpath.sdk.authc.AuthenticationResult;
@@ -22,20 +22,23 @@ import com.stormpath.sdk.servlet.http.Saver;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
+import org.springframework.security.web.DefaultSecurityFilterChain;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.security.web.csrf.CsrfTokenRepository;
-import org.springframework.stereotype.Component;
 
 /**
-* @since 1.0.RC5
-*/
-@Component
-public class StormpathWebSecurityConfigurer {
+ * @since 1.0.RC5
+ */
+@Configuration
+@EnableStormpathWebSecurity
+public class StormpathWebSecurityConfigurer extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
 
     @Autowired
     protected Client client;
@@ -55,10 +58,6 @@ public class StormpathWebSecurityConfigurer {
     @Autowired
     @Qualifier("stormpathCsrfTokenManager")
     protected CsrfTokenManager csrfTokenManager;
-
-    @Autowired
-    @Qualifier("stormpathAuthenticationProvider")
-    protected AuthenticationProvider stormpathAuthenticationProvider; //provided by stormpath-spring-security
 
     @Autowired(required = false) //required = false when stormpath.web.enabled = false
     @Qualifier("stormpathAuthenticationResultSaver")
@@ -115,36 +114,55 @@ public class StormpathWebSecurityConfigurer {
     @Value("#{ @environment['stormpath.web.csrfProtection.enabled'] ?: true }")
     protected boolean csrfProtectionEnabled;
 
+    @Value("#{ @environment['stormpath.spring.security.fullyAuthenticated.enabled'] ?: true }")
+    protected boolean fullyAuthenticatedEnabled;
+
+    public static StormpathWebSecurityConfigurer stormpath() {
+        return new StormpathWebSecurityConfigurer();
+    }
+
     /**
      * The pre-defined Stormpath access control settings are defined here.
+     * <p>This method has been marked as final in order to avoid users to override this method by mistake and thus removing
+     * all the required configurations. Instead, users can extend this class and configure their applications by overriding
+     * the {@link #doConfigure(HttpSecurity)} method. This way the configuration can be explicitly modified but not overwritten
+     * by mistake.</p>
      *
-     * @param http the {@link HttpSecurity} to be modified
-     * @throws Exception if an error occurs
+     * @param http
+     *            the {@link HttpSecurity} to be modified
+     * @throws Exception
+     *             if an error occurs
+     * @see #doConfigure(HttpSecurity)
      */
-    public final void configure(HttpSecurity http) throws Exception {
+    @Override
+    public void init(HttpSecurity http) throws Exception {
+
+        // autowire this bean
+        ApplicationContext context = http.getSharedObject(ApplicationContext.class);
+        context.getAutowireCapableBeanFactory().autowireBean(this);
 
         if (stormpathWebEnabled) {
             if (loginEnabled) {
                 http
-                    .formLogin()
-                    .loginPage(loginUri)
-                    .defaultSuccessUrl(loginNextUri)
-                    .successHandler(successHandler)
-                    .usernameParameter("login")
-                    .passwordParameter("password")
-                    .and().authorizeRequests()
-                    .antMatchers(loginUri).permitAll();
+                        .formLogin()
+                        .loginPage(loginUri)
+                        .defaultSuccessUrl(loginNextUri)
+                        .successHandler(successHandler)
+                        .usernameParameter("login")
+                        .passwordParameter("password")
+                        .and().authorizeRequests()
+                        .antMatchers(loginUri+"*").permitAll();
             }
 
             if (logoutEnabled) {
                 http
-                    .logout()
-                    .invalidateHttpSession(true)
-                    .logoutUrl(logoutUri)
-                    .logoutSuccessUrl(logoutNextUri)
-                    .addLogoutHandler(logoutHandler)
-                    .and().authorizeRequests()
-                    .antMatchers(logoutUri).permitAll();
+                        .logout()
+                        .invalidateHttpSession(true)
+                        .logoutUrl(logoutUri)
+                        .logoutSuccessUrl(logoutNextUri)
+                        .addLogoutHandler(logoutHandler)
+                        .and().authorizeRequests()
+                        .antMatchers(logoutUri).permitAll();
 
             }
 
@@ -168,27 +186,65 @@ public class StormpathWebSecurityConfigurer {
                 http.authorizeRequests().antMatchers(verifyUri).permitAll();
             }
 
-            //http.authenticationProvider(stormpathAuthenticationProvider);
-        }
-    }
-
-    public void configure(WebSecurity web) throws Exception {
-        if (stormpathWebEnabled) {
-            web.ignoring()
-                .antMatchers("/assets/css/stormpath.css")
-                .antMatchers("/assets/css/custom.stormpath.css");
+            if (fullyAuthenticatedEnabled) {
+                http
+                    .authorizeRequests()
+                    .antMatchers("/assets/**").permitAll()
+                    .anyRequest().fullyAuthenticated();
+            }
         }
     }
 
     /**
-     * Method to specify the {@link AuthenticationProvider} that Spring Security will use when processing authentications.
+     * Convenience <a href="https://en.wikipedia.org/wiki/Template_method_pattern">Hook Method</a> that will be invoked
+     * by {@link #configure(HttpSecurity)} after configuring all the properties required by Stormpath. You can override
+     * this method to define app-specific security settings like:
      *
-     * @param auth the {@link AuthenticationManagerBuilder} to use
-     * @throws Exception if an error occurs
+     * <pre>
+     * http
+     *   .authorizeRequests()
+     *   .antMatchers("/account").fullyAuthenticated()
+     *   .antMatchers("/admin").hasRole("ADMIN");
+     * </pre>
+     *
+     * @param http
+     *            the {@link HttpSecurity} to modify
+     * @throws Exception
+     *             if an error occurs
      */
-    public void configure(AuthenticationManagerBuilder auth) throws Exception {
-        if (stormpathWebEnabled && stormpathSecuritybEnabled) {
-            auth.authenticationProvider(stormpathAuthenticationProvider);
-        }
+    @Deprecated
+    protected void doConfigure(HttpSecurity http) throws Exception {
+
     }
+
+    /**
+     * Convenience <a href="https://en.wikipedia.org/wiki/Template_method_pattern">Hook Method</a> that will be invoked
+     * by {@link #configure(AuthenticationManagerBuilder)} after configuring all the properties required by Stormpath. You can
+     * override this method to define app-specific ones.
+     *
+     * @param auth
+     *            the {@link AuthenticationManagerBuilder} to modify
+     * @throws Exception
+     *             if an error occurs
+     */
+    @Deprecated
+    protected void doConfigure(AuthenticationManagerBuilder auth) throws Exception {
+
+    }
+
+    /**
+     * Convenience <a href="https://en.wikipedia.org/wiki/Template_method_pattern">Hook Method</a> that will be invoked
+     * by {@link #configure(WebSecurity)} after configuring all the properties required by Stormpath. You can override
+     * this method to define app-specific ones.
+     *
+     * @param web
+     *            the {@link WebSecurity} to modify
+     * @throws Exception
+     *             if an error occurs
+     */
+    @Deprecated
+    protected void doConfigure(WebSecurity web) throws Exception {
+
+    }
+
 }
