+++ tutorials/spring-boot/04-a-finer-grain-of-control/src/main/java/com/stormpath/tutorial/config/SpringSecurityWebAppConfig.java
@@ -15,7 +15,7 @@
  */
 package com.stormpath.tutorial.config;
 
-import com.stormpath.spring.config.StormpathWebSecurityConfigurerAdapter;
+import com.stormpath.spring.config.StormpathWebSecurityConfigurer;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 
@@ -23,7 +23,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
  * @since 1.0.RC5
  */
 @Configuration
-public class SpringSecurityWebAppConfig extends StormpathWebSecurityConfigurerAdapter {
+public class SpringSecurityWebAppConfig extends StormpathWebSecurityConfigurer {
 
     @Override
     protected void doConfigure(HttpSecurity http) throws Exception {
